Move ArrayData to ArrayData.cs
